vpn_backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.7
django-celery-beat==2.8.0
django-cors-headers==4.3.0
django-timezone-field==7.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
idna==3.10
inflection==0.5.1
kombu==5.5.3
packaging==24.2
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyroute2==0.7.12
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.31.0
segno==1.6.6
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
win-inet-pton==1.1.0
wireguard-tools==0.4.2
